+++ example/sample_flutter_app/lib/computer/computer.dart
@@ -11,7 +11,10 @@ class _GeneratedExamplesState extends State<GeneratedExamples>{
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      body: _buildBody(),
+      // body: _buildBody()
+      body: SingleChildScrollView(
+        child: _buildBody(),
+      ),
     );
   }
   Widget _buildBody(){