+++ figma/src/app/components/highlight.tsx
@@ -8,7 +8,7 @@ import "./highlight.css"
 import Prism from 'prism-react-renderer/prism';
 import dartLang from 'refractor/lang/dart';
 import { EK_COPIED } from '../constants/ek.constant';
-import { quickLook } from '../../dev-tools/quicklook';
+import { quickLook } from '../../quicklook';
 import { Widget } from '@bridged.xyz/flutter-builder';
 import { notify } from '@bridged.xyz/design-sdk/lib/figma';
 dartLang(Prism);
@@ -52,6 +52,7 @@ export default class Highlight extends React.Component<Props, State> {
             setLoadingState(false)
             notify(parent, "quick look ready !")
         }).catch((e) => {
+            console.error(e)
             setLoadingState(false)
             notify(parent, "compile failed. view console for details.", 2)
         })