+++ packages/detection/lib/list.detection/index.ts
@@ -1,3 +1,0 @@
-export function detectIfList(node: SceneNode) {
-
-}